Renamed kit package to internal
