+++ tests/conformance/bindings/bindings.go
@@ -14,7 +14,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/dapr/components-contrib/bindings"
-	"github.com/dapr/components-contrib/kit/config"
+	"github.com/dapr/components-contrib/internal/config"
 	"github.com/dapr/components-contrib/tests/conformance/utils"
 	"github.com/dapr/dapr/pkg/logger"
 )
